requirements.txt
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
covid-data-api==1.5
Flask==1.1.2
Flask-Cors==3.0.9
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pydantic==1.6.1
python-dateutil==2.8.1
requests==2.23.0
six==1.15.0
soupsieve==2.0.1
typer==0.3.2
urllib3==1.25.10
Werkzeug==1.0.1
